Fix openssl 3 and curl 7 variation configuration
There was a typo and curl 7 was not being build linked against openssl3

We also switch from curl to libcurl because the actual depedency is on
libcurl. libcurl also is the only one that has run_exports.

See a great analysis of this issue in the tiledb-feedstock: https://github.com/conda-forge/tiledb-feedstock/issues/224